pyproject.toml
[tool.poetry]
name = "gym-rat"
version = "0.1.0"
description = ""
authors = ["Kip Brown <krb24126@gmail.com>"]
readme = "README.md"
packages = [
    { include="backend", from="."}
]

[tool.poetry.scripts]
gymrat = "backend.cli.main:app"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.85.1"
uvicorn = "0.19.0"
sqlmodel = "0.0.8"
typer = "0.6.1"
rich = "12.6.0"
loguru = "0.6.0"
httpx = "0.23.0"
Jinja2 = "^3.1.2"
python-jose = "3.3.0"
passlib = "1.7.4"
bcrypt = "4.0.1"
pendulum = "2.1.2"
pydantic = {version = "1.10.2", extras = ["email"]}
pg8000 = "1.29.3"
python-multipart = "0.0.5"

[tool.poetry.dev-dependencies]
pytest = "7.2.0"
pytest-sugar = "0.9.5"
pytest-randomly = "3.12.0"
pytest-asyncio = "0.20.1"
ipython = "^8.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
